Added too much for one commit.  See notes below.
I renamed ImageSource to FileSource and added a separate class to make a
real "ImageSource" (i.e. one that sources from a list of numpy arrays.)

I also added a bunch of (as-yet) untested methods to the hopperchain class
that will help when it comes time to make the user interface.

Additionally, I moved the TEST_IMAGES out into a separate nose_const module
that I'll use for test-related constants from now on.